доделана аутенфикация, добавлены Policy для методов

+++ backend/app/Http/Controllers/api/v1/SnippetController.php
@@ -7,10 +7,9 @@ use App\Http\Requests\StoreSnippetRequest;
 use App\Http\Requests\UpdateSnippetRequest;
 use App\Http\Resources\api\v1\Snippet\SnippetCollection;
 use App\Http\Resources\api\v1\Snippet\SnippetResource;
-use App\Models\Snippet;
+use App\Models\User;
 use App\Services\SnippetService;
-use Hashids\Hashids;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\Request;
 
 class SnippetController extends Controller
 {
@@ -21,16 +20,23 @@ class SnippetController extends Controller
         $this->snippetService = $snippetService;
     }
 
-    public function index()
+    public function index(User $user)
     {
-        $user = auth()->user();
+        return new SnippetCollection($user->snippets);
+    }
+
+    public function indexOfAuthUser()
+    {
+        $user = auth('sanctum')->user();
         return new SnippetCollection($user->snippets);
     }
 
     public function store(StoreSnippetRequest $request)
     {
         $data = $request->validated();
-
+        if (auth('sanctum')->check()) {
+            $data['user_id'] = auth('sanctum')->user()->id;
+        }
         return new SnippetResource($this->snippetService->createSnippet($data));
     }
 
@@ -41,12 +47,18 @@ class SnippetController extends Controller
 
     public function update(UpdateSnippetRequest $request, $unique_id)
     {
+        if (!$this->snippetService->validate($request, $unique_id, 'update')) {
+            return response()->json(['message' => 'Forbidden'], 403);
+        }
         $data = $request->validated();
         return new SnippetResource($this->snippetService->updateSnippet($data, $unique_id));
     }
 
-    public function destroy($unique_id)
+    public function destroy(Request $request, $unique_id)
     {
+        if (!$this->snippetService->validate($request, $unique_id, 'delete')) {
+            return response()->json(['message' => 'Forbidden'], 403);
+        }
         return $this->snippetService->deleteSnippet($unique_id);
     }
 }
